checkAuthorization

+++ src/helpers/createMutation.js
@@ -1,34 +1,68 @@
 // @flow
 
-export function createRemoveMutation({ Model, populate = [] }: {
+export function createRemoveMutation({
+  Model,
+  populate = [],
+  checkAuthorization,
+}: {
   Model: any,
   populate: Array<string>,
+  checkAuthorization?: (context: any) => Promise<any>,
 }) {
-  return (parentObj: any, product: TProduct) => {
-    const { _id } = product
+  return async (parentObj: any, entity: any, context: any) => {
+    if (checkAuthorization) {
+      const error = await checkAuthorization(context)
+      if (error) return { error }
+    }
+    const { _id } = entity
     const q = Model.findOneAndRemove({ _id })
     populate.forEach(field => q.populate(field))
-    return q.lean()
+    const result = await q.lean()
+    return {
+      result,
+    }
   }
 }
 
-export function createCreateMutation({ Model, populate = [] }: {
+export function createCreateMutation({
+  Model,
+  populate = [],
+  checkAuthorization,
+}: {
   Model: any,
   populate: Array<string>,
+  checkAuthorization?: (context: any) => Promise<any>,
 }) {
-  return async (parentObj: any, product: TProduct) => {
-    const { _id } = await Model.create(product)
+  return async (parentObj: any, entity: any, context: any) => {
+    if (checkAuthorization) {
+      const error = await checkAuthorization(context)
+      if (error) return { error }
+    }
+    const { _id } = await Model.create(entity)
     const q = Model.findOne({ _id })
     populate.forEach(field => q.populate(field))
-    return q.lean()
+    const result = await q.lean()
+    return {
+      result,
+    }
   }
 }
 
-export function createEditMutation({ Model, populate = [] }: {
+export function createEditMutation({
+  Model,
+  populate = [],
+  checkAuthorization,
+}: {
   Model: any,
   populate: Array<string>,
+  checkAuthorization?: (context: any) => Promise<any>,
 }) {
-  return (parentObj: any, entity: any) => {
+  return async (parentObj: any, entity: any, context: any) => {
+    if (checkAuthorization) {
+      const error = await checkAuthorization(context)
+      if (error) return { error }
+    }
+
     const { _id, ...rest } = entity
     const q = Model.findOneAndUpdate(
       { _id },
@@ -36,7 +70,10 @@ export function createEditMutation({ Model, populate = [] }: {
       { new: true },
     )
     populate.forEach(field => q.populate(field))
-    return q
+    const result = await q.lean()
+    return {
+      result,
+    }
   }
 }
 
@@ -44,14 +81,16 @@ export default function createMutation({
   name,
   Model,
   populate = [],
+  checkAuthorization,
 }: {
   Model: any,
   name: string,
   populate: Array<string>,
+  checkAuthorization?: (context: any) => Promise<any>,
 }) {
   return {
-    [`create${name}`]: createCreateMutation({ Model, populate }),
-    [`edit${name}`]: createEditMutation({ Model, populate }),
-    [`remove${name}`]: createRemoveMutation({ Model, populate }),
+    [`create${name}`]: createCreateMutation({ Model, populate, checkAuthorization }),
+    [`edit${name}`]: createEditMutation({ Model, populate, checkAuthorization }),
+    [`remove${name}`]: createRemoveMutation({ Model, populate, checkAuthorization }),
   }
 }
